+++ console/src/core/PluginResolver.ts
@@ -1,12 +1,23 @@
-import type { PluginList } from "@/extension/ArgumentPlugin";
+import type {PluginCommand, PluginDetail, PluginList, PluginToolbar} from "@/extension/ArgumentPlugin";
 import type { AbstractPlugin } from "@/extension/BasePlugin";
+import {Stack} from "@/core/BasicStructure";
+import type {ToolbarItemOptions} from "@toast-ui/editor/types/ui";
+import type {LucenceCore} from "@/core/LucenceCore";
+import {DefaultPlugin} from "@/plugin/DefaultPlugin";
+import {TestPlugin} from "@/plugin/TestPlugin";
 
 export class PluginResolver {
     
-    private _pluginList: PluginList = [];
+    private _pluginList: PluginList = new Stack<PluginDetail>();
     
-    // 装配Plugin到内存
-    public assemble(plugin: AbstractPlugin): void {
+    private readonly core: LucenceCore;
+    
+    constructor(core: LucenceCore) {
+        this.core = core;
+    }
+
+    // 装配Plugin
+    private load(plugin: AbstractPlugin): void {
         // detail数据校验
         if (!plugin.detail) {
             throw Error("Plugin doesn't have detail.");
@@ -27,8 +38,64 @@ export class PluginResolver {
         if (!plugin.detail.github || !matchUrl(plugin.detail.github)) {
             throw Error("Plugin must provide a correct github url.");
         }
-        // 压入plugin列表内
+        // TODO 是否已经存在相同name和display的插件，如果是则不load这个插件并throw异常
+        
+        // 压栈
         this._pluginList.push(plugin.detail);
+        plugin.onEnable();
+
+        const commands: PluginCommand[] | null = plugin.createCommands();
+        // 初始化插件的commands
+        if (commands) {
+            commands.forEach((cmd: PluginCommand): void => {
+                this.core.editor.addCommand(
+                    'markdown',
+                    cmd.name,
+                    cmd.command,
+                )
+            });
+        }
+        // 初始化插件的toolbar
+        const toolbar: PluginToolbar | null = plugin.createToolbar();
+        if (toolbar) {
+            const items: ToolbarItemOptions[] = toolbar.items;
+            for (let i: number = 0; i < items.length; i++) {
+                this.core.editor.insertToolbarItem({
+                    groupIndex: 0,
+                    itemIndex: i,
+                }, items[i])
+            }
+        }
+    }
+    
+    // 卸载Plugin
+    public disable(plugin: AbstractPlugin): void {
+        // 弹栈
+        plugin.onDisable();
+    }
+    
+    // 自动挂载
+    public autoload(): void {
+        // 硬加载基本插件
+        const defaultOne: DefaultPlugin = new DefaultPlugin(this.core);
+        this.load(defaultOne);
+        const testOne: TestPlugin = new TestPlugin(this.core);
+        this.load(testOne);
+        // 扫描
+        this.scanAll();
+        
+    }
+    
+    // 扫描所有plugin并通过load挂载
+    private scanAll() {
+        
+    }
+
+    /**
+     * 获取plugin列表
+     */
+    public get pluginList() {
+        return this._pluginList;
     }
     
 }
