+++ console/src/css/EditorStyle.css
@@ -20,7 +20,7 @@
     --popup-hover-bg: #f2f2f2;
     --line-highlight: #000;
     --code-normal-color: #6e7681;
-    --code-normal-bg: hsl(230, 1%, 98%);
+    --code-normal-bg: #fff;
     --code-amber-popup: #fff;
     --code-amber-shadow: rgba(0, 0, 0, 0.16) 0 0 8px 2px;
     --code-amber-search-border: rgba(0, 0, 0, .08);
@@ -29,6 +29,10 @@
     --code-amber-input-border: rgba(0, 0, 0, .12);
     --code-amber-btn-disable: rgba(97, 97, 97, 0.5);
     --code-amber-block-bg: rgba(0, 0, 0, .03);
+    --plugin-bg: #fff;
+    --plugin-card--active: rgb(228,230,241);
+    --plugin-column--border: 1px solid rgba(0, 0, 0, .12);
+    --plugin-font--sub: rgba(0,0,0,.6);
 }
 
 [data-theme="night"] {
@@ -62,6 +66,10 @@
     --code-amber-input-border: rgba(255, 255, 255, .12);
     --code-amber-btn-disable: rgba(204, 204, 204, 0.5);
     --code-amber-block-bg: rgba(255, 255, 255, .12);
+    --plugin-bg: rgb(32,33,35);
+    --plugin-card--active: rgb(4,57,94);
+    --plugin-column--border: 1px solid rgb(43,43,43);
+    --plugin-font--sub: rgba(255,255,255,.6);
 }
 
 .lucence-wrapper {
@@ -175,7 +183,10 @@ span.stat-panel--key.last { margin-right: -10px; }
 }
 
 .toastui-editor-toolbar-group-tail { display: flex; }
-.toastui-editor-defaultUI-toolbar .toastui-editor-toolbar-group { flex-grow: 1; }
+.toastui-editor-defaultUI-toolbar .toastui-editor-toolbar-group { 
+    flex-grow: 1;
+    align-items: center;
+}
 .toastui-editor-defaultUI-toolbar button {
     color: var(--editor-toolbar-btn);
     border: none!important;
@@ -717,16 +728,18 @@ pre.hljs,
     height: 100%;
     z-index: 999;
     background: rgba(86,88,105,.7);
+    display: flex;
+    align-items: center;
 }
 
 #lucence-plugin--store .lucence-plugin--container {
     width: 80%;
     height: 80%;
     padding: 20px;
-    margin: 8% auto;
+    margin: 0 auto;
     font-size: 20px;
-    border-radius: 5px;
-    background: rgb(32,33,35);
+    border-radius: 8px;
+    background: var(--plugin-bg);
 }
 #lucence-plugin--store .lucence-plugin--container .lucence-plugin--head {
     display: flex;
@@ -744,6 +757,12 @@ pre.hljs,
 .lucence-plugin--head .plugin-head--title {
     justify-content: start;
     text-align: start;
+}
+.lucence-plugin--head .plugin-head--title span {
+    font-size: 90%;
+    font-weight: normal;
+    color: #999;
+    margin-left: 10px;
 }
 .lucence-plugin--head .plugin-head--close {
     justify-content: end;
@@ -769,12 +788,50 @@ pre.hljs,
 }
 .lucence-plugin--body .lucence-plugin--list {
     flex: 0.3;
-    border: 1px solid rgb(43, 43, 43);
+    border: var(--plugin-column--border);
     border-radius: 5px;
 }
 .lucence-plugin--body .lucence-plugin--detail {
     flex: 0.7;
     margin-left: 10px;
-    border: 1px solid rgb(43, 43, 43);
+    border: var(--plugin-column--border);
     border-radius: 5px;
+}
+.lucence-plugin--card:first-child {
+    border-radius: 5px 5px 0 0;
+}
+.lucence-plugin--card:last-child {
+    border-radius: 0 0 5px 5px;
+}
+.lucence-plugin--card {
+    display: flex;
+    width: 100%;
+    height: 84px;
+    padding: 10px;
+    cursor: pointer;
+}
+.lucence-plugin--card:hover { background: var(--popup-hover-bg); }
+.lucence-plugin--card.active { 
+    cursor: default!important;
+    background: var(--default-highlight)!important; 
+}
+.lucence-plugin--card .left-column { height: 100%; width: 64px; }
+.lucence-plugin--card .left-column img { border-radius: 3px; }
+.lucence-plugin--card .right-column {
+    flex: 1;
+    height: 100%;
+    padding-left: 8px;
+    position: relative;
+}
+.lucence-plugin--card .right-column p { color: var(--display-p); }
+.lucence-plugin--card .right-column .plugin-info--title {
+    line-height: 22px;
+    font-weight: bold;
+}
+.lucence-plugin--card .right-column .plugin-info--simple {
+    position: absolute;
+    line-height: 20px;
+    font-size: 14px;
+    bottom: 0;
+    color: var(--plugin-font--sub);
 }
